adds functionality...finally
