+++ cypress/integration/user_can_choose_a_weapon.feature.js
@@ -2,8 +2,8 @@ describe('User can choose a weapon', () => {
     it('Selects a weapon from the menu', () => {
     cy.visit('/');
     cy.get('button#start').click();
-    cy.contains('Weapons').click();
-    cy.contains('#rock').click('left','Rock');
-    cy.get('img').should('contain', '#rock')
+    cy.contains('Weapon').click();
+    cy.contains('Prompt').type('rock');
+    cy.get('p').should('contain', '#rock')
   })
 })
